pyproject.toml
[project]
name = "financial-document-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "html-to-markdown>=2.14.11",
    "jinja2>=3.1.6",
    "jsonlines>=4.0.0",
    "langgraph>=1.0.1",
    "nltk>=3.9.2",
    "numpy>=2.3.4",
    "ollama>=0.6.0",
    "onnxruntime>=1.23.2",
    "pdf2image>=1.17.0",
    "pikepdf>=10.0.0",
    "pypdf>=6.1.3",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.51.0",
    "tiktoken>=0.12.0",
    "torch>=2.9.0",
    "tqdm>=4.67.1",
    "unstructured-inference>=1.0.5",
    "unstructured[local-inference,pdf]>=0.18.15",
    "weaviate-client>=4.17.0",
]
